rename recovery key dialog component
and related fields/properties
