+++ frontend/src/components/DisplayRecoveryKeyDialog.vue
